+++ storing/service_test.go
@@ -158,7 +158,7 @@ func TestCreateFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		defer server.Close()
 		remoteClient, err := client.New(server.URL)
@@ -196,7 +196,7 @@ func TestCreateFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		defer server.Close()
 		remoteClient, err := client.New(server.URL)
@@ -255,7 +255,7 @@ func TestGetFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		c, err := client.New(server.URL)
 		require.NoError(t, err)
@@ -288,7 +288,7 @@ func TestGetFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		defer server.Close()
 		c, err := client.New(server.URL)
@@ -347,7 +347,7 @@ func TestDeleteFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		c, err := client.New(server.URL)
 		require.NoError(t, err)
@@ -377,7 +377,7 @@ func TestDeleteFile(t *testing.T) {
 		m := mock.NewMembership(ctrl)
 		defer ctrl.Finish()
 
-		h := httptransport.MakeHandler(s2, &config.Config{})
+		h := httptransport.MakeHandler(s2, &config.Config{}, func() bool { return true })
 		server := httptest.NewServer(h)
 		defer server.Close()
 		c, err := client.New(server.URL)